refactor: generate notion + dify commands from the factory
Migrate the read-only SaaS backends notion and dify to make_generic_commands
(gmail/gdrive pattern). notion keeps only the bespoke notion_* writers; dify
keeps cat/find (avoid an extra document-detail API call), search (retrieval
pushdown), and sed (no generic builder).

Also fix the generic stat builder to thread the cache index into generic_stat,
so index-backed backends resolve directory mtime on a cold call (notion stat
-c '%y' returned empty before).
