+++ python/mirage/commands/builtin/dify/__init__.py
@@ -1,35 +1,61 @@
-from mirage.commands.builtin.dify.awk import awk
+# ========= Copyright 2026 @ Strukto.AI All Rights Reserved. =========
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+# ========= Copyright 2026 @ Strukto.AI All Rights Reserved. =========
+
 from mirage.commands.builtin.dify.cat import cat
-from mirage.commands.builtin.dify.cut import cut
 from mirage.commands.builtin.dify.find import find
-from mirage.commands.builtin.dify.grep import grep
-from mirage.commands.builtin.dify.head import head
-from mirage.commands.builtin.dify.ls import ls
-from mirage.commands.builtin.dify.rg import rg
 from mirage.commands.builtin.dify.search import search
 from mirage.commands.builtin.dify.sed import sed
-from mirage.commands.builtin.dify.sort import sort
-from mirage.commands.builtin.dify.stat import stat
-from mirage.commands.builtin.dify.tail import tail
-from mirage.commands.builtin.dify.tree import tree
-from mirage.commands.builtin.dify.uniq import uniq
-from mirage.commands.builtin.dify.wc import wc
+from mirage.commands.builtin.generic_bind import (CommandIO,
+                                                  make_generic_commands)
+from mirage.commands.builtin.generic_bind.provision import (
+    make_search_provision, metadata_provision)
+from mirage.core.dify.read import read_bytes as _read
+from mirage.core.dify.read import read_stream as _read_stream
+from mirage.core.dify.readdir import readdir as _readdir
+from mirage.core.dify.stat import stat as _stat
+
+# Dify knowledge-base documents are read through the generic factory. cat and
+# find keep wrappers to avoid an extra document-detail API call per path: the
+# generic cat eagerly stats the file and the generic find would call the full
+# stat, while Dify's stat is a detail fetch (the wrappers use bespoke glob
+# resolution / stat_light instead). search pushes down to the Dify retrieval
+# API and sed has no generic builder. Dify is read-only, so the generic
+# byte-mutation commands are intentionally absent (no write op wired).
+_DIFY_CMD_OPS = CommandIO(
+    readdir=_readdir,
+    read_bytes=_read,
+    read_stream=_read_stream,
+    stat=_stat,
+    is_mounted=lambda a: True,
+    local=False,
+)
+
+_DIFY_OVERRIDES = {"cat", "find"}
 
 COMMANDS = [
-    awk,
+    *make_generic_commands(
+        "dify",
+        _DIFY_CMD_OPS,
+        overrides=_DIFY_OVERRIDES,
+        provision_overrides={
+            "grep": make_search_provision(_stat),
+            "rg": make_search_provision(_stat),
+            "ls": metadata_provision,
+        },
+    ),
     cat,
-    cut,
     find,
-    grep,
-    head,
-    ls,
-    rg,
     search,
     sed,
-    sort,
-    stat,
-    tail,
-    tree,
-    uniq,
-    wc,
 ]
