--- python/mirage/commands/builtin/notion/ls.py
@@ -1,77 +1,0 @@
-# ========= Copyright 2026 @ Strukto.AI All Rights Reserved. =========
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-# ========= Copyright 2026 @ Strukto.AI All Rights Reserved. =========
-
-from functools import partial
-
-from mirage.accessor.notion import NotionAccessor
-from mirage.cache.index import IndexCacheStore
-from mirage.commands.builtin.generic.ls import ls as generic_ls
-from mirage.commands.builtin.notion._provision import metadata_provision
-from mirage.commands.registry import command
-from mirage.commands.spec import SPECS
-from mirage.core.notion.glob import resolve_glob
-from mirage.core.notion.readdir import readdir
-from mirage.core.notion.stat import stat
-from mirage.io.types import ByteSource, IOResult
-from mirage.provision.types import ProvisionResult
-from mirage.types import LsSortBy, PathSpec
-
-
-async def ls_provision(
-    accessor: NotionAccessor,
-    paths: list[PathSpec],
-    *texts: str,
-    **_extra: object,
-) -> ProvisionResult:
-    return await metadata_provision("ls " + " ".join(
-        p.original if isinstance(p, PathSpec) else p for p in paths))
-
-
-@command("ls", resource="notion", spec=SPECS["ls"], provision=ls_provision)
-async def ls(
-    accessor: NotionAccessor,
-    paths: list[PathSpec],
-    *texts: str,
-    stdin: bytes | None = None,
-    args_l: bool = False,
-    args_1: bool = False,
-    a: bool = False,
-    A: bool = False,
-    h: bool = False,
-    t: bool = False,
-    S: bool = False,
-    r: bool = False,
-    R: bool = False,
-    d: bool = False,
-    F: bool = False,
-    index: IndexCacheStore = None,
-    **_extra: object,
-) -> tuple[ByteSource | None, IOResult]:
-    paths = await resolve_glob(accessor, paths, index)
-    sort_by = LsSortBy.TIME if t else LsSortBy.SIZE if S else LsSortBy.NAME
-    return await generic_ls(
-        paths,
-        readdir=partial(readdir, accessor),
-        stat=partial(stat, accessor),
-        long=args_l,
-        one_per_line=args_1,
-        all_files=a or A,
-        human=h,
-        sort_by=sort_by,
-        reverse=r,
-        recursive=R,
-        list_dir=d,
-        classify=F,
-        index=index,
-    )
