--- python/mirage/commands/builtin/notion/realpath.py
@@ -1,47 +1,0 @@
-# ========= Copyright 2026 @ Strukto.AI All Rights Reserved. =========
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-# ========= Copyright 2026 @ Strukto.AI All Rights Reserved. =========
-
-from functools import partial
-
-from mirage.accessor.notion import NotionAccessor
-from mirage.cache.index import IndexCacheStore
-from mirage.commands.builtin.generic.realpath import \
-    realpath as generic_realpath
-from mirage.commands.registry import command
-from mirage.commands.spec import SPECS
-from mirage.core.notion.glob import resolve_glob
-from mirage.core.notion.stat import stat as stat_impl
-from mirage.io.types import ByteSource, IOResult
-from mirage.types import PathSpec
-
-
-@command("realpath", resource="notion", spec=SPECS["realpath"])
-async def realpath(
-    accessor: NotionAccessor,
-    paths: list[PathSpec] | None = None,
-    *texts: str,
-    stdin: bytes | None = None,
-    e: bool = False,
-    m: bool = False,
-    prefix: str = "",
-    index: IndexCacheStore = None,
-    **_extra: object,
-) -> tuple[ByteSource | None, IOResult]:
-    if paths:
-        paths = await resolve_glob(accessor, paths, index)
-    return await generic_realpath(paths or [],
-                                  stat_fn=partial(stat_impl, index=index),
-                                  accessor=accessor,
-                                  e=e,
-                                  m=m)
